Move SignatureHash to Transaction namespace
